Example of "Printf" C AVR library
Library allows you to use printf and sprintf in C code to output a value to a serial port.  I mostly use this to debug simple programs when I don't think using debugwire or a simulator is warranted.  the Example uses this library to print LED status each time an LED is flashed to a PC terminal.  Assumption hardware wise is an arduino uno with 16Mhz 328P.